changed int to double

+++ include/videostream.h
@@ -6,9 +6,9 @@
 
 int setup_capture(const int device);
 
-IplImage* grab_image(const int scale, const int convert_grayscale);
+IplImage* grab_image(const double scale, const int convert_grayscale);
 
-unsigned char* grab_raw_data(const int scale, const int convert_grayscale, int* imagesize);
+unsigned char* grab_raw_data(const double scale, const int convert_grayscale, int* imagesize);
 
 int clean_up_stream(void);
